requirements.txt
openai-whisper==20231117
ollama
yt-dlp
